(fix): ensure single refresh token call on concurrent 401 errors

+++ src/app/core/interceptors/token.interceptor.ts
@@ -1,36 +1,58 @@
 import { HttpInterceptorFn } from '@angular/common/http';
 import { inject } from '@angular/core';
-import { catchError, switchMap, throwError } from 'rxjs';
+import {
+  BehaviorSubject,
+  catchError,
+  filter,
+  switchMap,
+  take,
+  throwError,
+} from 'rxjs';
 import { AccountService } from '../services/account.service';
 import { Router } from '@angular/router';
 
+let isRefreshing = false;
+// save state of token refresh
+const refreshTokenSubject = new BehaviorSubject<boolean>(false);
+
 export const tokenInterceptor: HttpInterceptorFn = (req, next) => {
   const accountService = inject(AccountService);
-  const route = inject(Router);
+  const router = inject(Router);
 
   return next(req.clone({ withCredentials: true })).pipe(
     catchError((error) => {
       if (error.status === 401) {
-        // Try to refresh the token
-        return accountService.refreshToken().pipe(
-          // Switch to retrying the original request after successful token refresh
-          switchMap(() => {
-            // Create a new request with the same configuration as the original
-            return next(req.clone({ withCredentials: true }));
-          }),
-          // Handle errors during token refresh
-          catchError((refreshError) => {
-            console.error('Error refreshing token:', refreshError);
-            // Logout the user and redirect to login page
-            accountService.removeCurrentUserSource();
-            route.navigateByUrl('/login');
-            return throwError(
-              () => new Error('Session expired. Please login again.')
-            );
-          })
-        );
-      }
+        if (!isRefreshing) {
+          isRefreshing = true;
+          refreshTokenSubject.next(false);
 
+          return accountService.refreshToken().pipe(
+            switchMap(() => {
+              isRefreshing = false;
+              refreshTokenSubject.next(true);
+              // Retry the original request
+              return next(req.clone({ withCredentials: true }));
+            }),
+            // Handle errors during token refresh
+            catchError((refreshError) => {
+              // Logout the user and redirect to login page
+              isRefreshing = false;
+              accountService.removeCurrentUserSource();
+              router.navigateByUrl('/login');
+              return throwError(
+                () => new Error('Session expired. Please login again.')
+              );
+            })
+          );
+        } else {
+          // Wait until refresh completes
+          return refreshTokenSubject.pipe(
+            filter((refreshed) => refreshed === true),
+            take(1),
+            switchMap(() => next(req.clone({ withCredentials: true })))
+          );
+        }
+      }
       return throwError(() => error);
     })
   );
